requirements.txt
flask
mysql-connector-python
redis
Flask-Session
bcrypt
statsmodels
matplotlib
pandas